+++ src/Web/TizianaTerenzi.Web.ViewModels/DiscountCodes/CreateDiscountCodeInputModel.cs
@@ -8,8 +8,8 @@
         private const int NameMinimumLength = 3;
         private const int NameMaximumLength = 30;
 
-        private const double MinimumDiscount = 1.00;
-        private const double MaximumDiscount = 100.00;
+        private const int MinimumDiscount = 1;
+        private const int MaximumDiscount = 100;
 
         private const string NameErrorMessage = "Discount code should be at least 3 characters long and not more than 30.";
         private const string DiscountErrorMessage = "Discount should be at least 1% and no more than 100%.";
@@ -20,7 +20,7 @@
 
         [Required(ErrorMessage = "Discount is required.")]
         [Range(MinimumDiscount, MaximumDiscount, ErrorMessage = DiscountErrorMessage)]
-        public double Discount { get; set; }
+        public byte Discount { get; set; }
 
         public DateTime? ExpiresOn { get; set; }
     }
